DocuComp/app/components/AdvancedSearch.tsx
import React, { useState, ChangeEvent } from 'react';

interface SearchParams {
    title: string;
    category: string;
    dateFrom: string;
    dateTo: string;
    status: string;
}

const AdvancedSearch = () => {
    const [searchParams, setSearchParams] = useState<SearchParams>({
        title: '',
        category: '',
        dateFrom: '',
        dateTo: '',
        status: ''
    });
    const [searchResults, setSearchResults] = useState<{ title: string }[]>([]);

    const handleInputChange = (e: ChangeEvent<HTMLInputElement | HTMLSelectElement>) => {
        const { name, value } = e.target;
        setSearchParams(prev => ({ ...prev, [name]: value }));
    };

    const executeSearch = async () => {
        // Mocked fetch function to simulate fetching data
        const fetchedData = [{ title: 'Document 1' }, { title: 'Document 2' }]; // Example data
        const error = false; // Simulate no error

        if (error) {
            console.error('Error fetching documents:', error);
        } else {
            setSearchResults(fetchedData);
        }
    };

    return (
        <div>
            <input
                type="text"
                name="title"
                value={searchParams.title}
                onChange={handleInputChange}
                placeholder="Document Title"
            />
            <select
                name="category"
                value={searchParams.category}
                onChange={handleInputChange}
            >
                <option value="">Select a Category</option>
                <option value="legal">Legal</option>
                <option value="finance">Finance</option>
            </select>
            {/* Add other input fields similarly */}
            <button onClick={executeSearch}>Search</button>
            <div>
                {searchResults.map((result, index) => (
                    <div key={index}>
                        <p>{result.title}</p>
                    </div>
                ))}
            </div>
        </div>
    );
};

export default AdvancedSearch;
